Improving API readability

+++ julez-core/src/main/java/uk/co/acuminous/julez/scenario/source/SizeLimiter.java
@@ -6,9 +6,9 @@ import uk.co.acuminous.julez.scenario.ScenarioSource;
 public class SizeLimiter implements ScenarioSource {
 
     private final ScenarioSource scenarios;
-    private int count;
+    private long count;
 
-    public SizeLimiter(ScenarioSource scenarios, int sizeLimit) {
+    public SizeLimiter(ScenarioSource scenarios, long sizeLimit) {
         this.scenarios = scenarios;        
         this.count = sizeLimit;
     }
